update: post merge cleanup
